% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bbox}
\alias{bbox}
\title{set up a bounding box for ggmap::get_map()}
\usage{
bbox(bb)
}
\arguments{
\item{bb}{boundings box specifications}
}
\description{
Converts a google bounding box to a stamen boundings box
}
