% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{bbox_lv95}
\alias{bbox_lv95}
\title{bbox in WSG84 from LV95}
\usage{
bbox_lv95(
  x = NA,
  y = NA,
  r = NA,
  w = NA,
  h = NA,
  x1 = NA,
  y1 = NA,
  x2 = NA,
  y2 = NA
)
}
\arguments{
\item{x}{LV95 x coordinate of center in m}

\item{y}{LV95 y coordinate of center in m}

\item{r}{radius around center in m (defines a quadrat)}

\item{w}{width of box in m (defines a rectancle)}

\item{h}{height of box in m (defines a rectancle)}

\item{x1}{left side (lower left point together with y1) in m}

\item{y1}{bottom side (lower left point together with x1) in m}

\item{x2}{right side (upper right point together with x2) in m}

\item{y2}{top side (upper right point together with y2) in m}
}
\value{
bbox as list with left, right, bottom, top, center = (lon, lat) as WSG84 coordinates, width and height in m
}
\description{
This Function creates a bbox for usage with \code{\link[=get_stamen_map]{get_stamen_map()}} or \code{\link[=get_google_map]{get_google_map()}}. Exactly one set of
(x, y, r) or (x, y, w, h) or (x1, y1, x2, y2) must be supplied as arguments. The actual position, width and height
in the resulting map can slightly differ on behalf of the projections from LV95 (EPSG:2056) to WSG84 (EPSG:4326) to
Pseudo Mercator (EPSG:3857). But should be negligible in most cases.
}
\examples{
# use in this form for background maps in polar plots
bbox_lv95(2683141, 1249040, 500)

# specific aspect ratio
bbox_lv95(2683141, 1249040, w = 700, h = 0.66 * 700)

# from a bounding box in LV95
bbox_lv95(x1 = 2683000, y1 = 1249000, x2 = 2683700, y2 = 1249500)
}
