% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcalendar.R
\name{cal_month_border}
\alias{cal_month_border}
\title{Add month border to calendar}
\usage{
cal_month_border(size = 1, lineend = "square", linejoin = "bevel",
  color = "grey5", ...)
}
\arguments{
\item{lineend}{Line end style (round, butt, square).}

\item{linejoin}{Line join style (round, mitre, bevel).}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\value{
ggplot2 layer
}
\description{
Add month border to calendar
}
