% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_cartpolar.R
\name{coord_cartpolar}
\alias{coord_cartpolar}
\title{cartesian coordinate system with polar grid lines}
\usage{
coord_cartpolar(limit = NULL, expand = TRUE, clip = "on",
  bg = NULL, scale_wd = NULL)
}
\arguments{
\item{limit}{limit for coordsystem (xlim(-limit, limit), ylim(-limit, limit))}

\item{expand}{expand limits (probably not used at the moment)}

\item{clip}{clip panel (not sure if used)}

\item{bg}{raster for background image}

\item{scale_wd}{pass breaks angle lines (not used at the moment)}
}
\description{
WIP:
\itemize{
\item render_axis_h / render_axis_v probably needs an overwrite
\item labels for polar grid lines
\item wd scale / breaks but how exactly?
}
}
\examples{
require(ggplot2)

df <- tibble::tribble(
  ~wd, ~ws,
  0, 4,
  90, 4,
  180, 4,
  270, 4,
  45, 4,
  135, 4,
  225, 4,
  315, 4,
  60, 9
)

df <- dplyr::mutate(df,
  u = .data$ws * sin(2 * pi * .data$wd / 360),
  v = .data$ws * cos(2 * pi * .data$wd / 360)
)

ggplot(df, aes(x=u, y=v)) + geom_point() + coord_cartpolar()

ggplot(df, aes(x=u, y=v)) + geom_point() + coord_cartpolar() +
  scale_x_continuous(breaks = c(1, 3, 4.5)) +
  theme("panel.grid.minor.x" = element_blank())


# background map
bbox <- tibble::tibble(x = c(2683141 - 500, 2683141 + 500), y = c(1249040 - 500, 1249040 + 500))
bbox <- rOstluft::transform_projection(bbox, coord = c("x", "y"),
                                       initCRS = sp::CRS("+init=epsg:2056"),
                                       outCRS = sp::CRS("+init=epsg:4326"))

bbox <- c(left = bbox$x[1], right = bbox$x[2], bottom = bbox$y[1], top = bbox$y[2])

raster_map <- ggmap::get_stamenmap(bbox, zoom = 16, maptype = "terrain",
                                   source = "stamen", color = "bw")

ggplot(df, aes(x=u, y=v)) + geom_point() + coord_cartpolar(bg = raster_map)
}
