% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_cartpolar2.R
\name{coord_cartpolar2}
\alias{coord_cartpolar2}
\title{cartesian coordinate system with polar grid lines}
\usage{
coord_cartpolar2(limit = NULL, clip = "on", bg = NULL,
  grill = c("background", "foreground"))
}
\arguments{
\item{limit}{limit for coordsystem (xlim(-limit, limit), ylim(-limit, limit))}

\item{clip}{clip panel (not sure if used)}

\item{bg}{raster for background image}

\item{expand}{expand limits}

\item{scale_wd}{pass breaks angle lines}

\item{labels_wd}{\code{\link[ggplot2:element_text]{ggplot2::element_text()}} object to format wd labels. Use \code{\link[ggplot2:element_blank]{ggplot2::element_blank()}}
to draw nothing}
}
\description{
A cartesian coordinate system overlayed with a polar grid.
It is used together with \code{\link[=summary_wind_2d]{summary_wind_2d()}} in \code{\link[=ggpolarplot]{ggpolarplot()}}.
}
\section{TODO}{

\itemize{
\item optimize scaling to match \code{\link[=coord_polar2]{coord_polar2()}} / \code{\link[ggplot2:polar]{ggplot2::polar()}}
\item Option to plot only positive y breaks
}
}

\examples{
library(ggplot2)

df <- tibble::tribble(
  ~wd, ~ws, ~facet,
  0, 5, 1,
  0, 2.5, 1,
  0, 9.5, 1,
  0, 4, 1,
  90, 4, 1,
  180, 4, 1,
  270, 4, 1,
  30, 8, 2,
  120, 8, 2,
  200, 8, 2,
  300, 8, 2
)

df <- dplyr::mutate(df,
  u = .data$ws * sin(2 * pi * .data$wd / 360),
  v = .data$ws * cos(2 * pi * .data$wd / 360)
)

p <-  ggplot(df, aes(x=u, y=v)) + geom_point()

p + coord_cartpolar2()


# y scale is used for uv scaling (negative breaks
# are ignored)
p  + coord_cartpolar2() +
  scale_y_continuous(
    limits = c(-10, 10),
    breaks = c(-5, 0, 1, 2.5, 7.5, 10),
    expand = c(0, 0, 0, 0)
  )

# alternative argument limit of coord_cartpolar2 can be used
p  + coord_cartpolar2(limit = 5)

# a continuous x scale is used to control the breaks and labels
# for wd. Setting limits in this scale will cause problems
p  + coord_cartpolar2() +
  scale_x_continuous(
    breaks = c(0, 45, 90, 180, 270),
    labels = c("Nord", "NE", "E", "S", "W")
  )


# theming: only major grid is draw
p + coord_cartpolar2() +
  theme(
    "panel.grid.major.x" = element_line(color = "red", size = 2, linetyp = "dashed"),
    "panel.grid.major.y" = element_line(color = "green", arrow = arrow()),
    "axis.text.x" = element_text(color = "violet", size = 14, face = "bold"),
    "axis.text.y" = element_text(color = "blue")
  )


# background map
bb <- bbox_lv95(2683141, 1249040, 500)
bg <- get_stamen_map(bb)

p + coord_cartpolar2(bg = bg)

# faceting
p + coord_cartpolar2(bg = bg) +
  facet_wrap(vars(facet))

# if plotting a raster layer, use grill = "foreground" to draw
# the polar grill over the raster and set the limit to max + 0.5
# and all expands to zero
raster <- expand.grid(u = -10:10, v = -10:10)
raster$z <- runif(nrow(raster))
ggplot(raster, aes(x=u, y=v, fill=z)) +
  coord_cartpolar2(bg = bg, grill = "foreground",  limit = 10.5) +
  geom_raster(alpha = 0.5) +
  scale_y_continuous(expand = c(0, 0, 0, 0))
}
