% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_polar2.R
\name{coord_polar2}
\alias{coord_polar2}
\title{Customized coord_polar}
\usage{
coord_polar2(theta = "x", start = 0, direction = 1, clip = "on")
}
\arguments{
\item{theta}{variable to map angle to (\code{x} or \code{y})}

\item{start}{offset of starting point from 12 o'clock in radians}

\item{direction}{1, clockwise; -1, anticlockwise}

\item{clip}{Should drawing be clipped to the extent of the plot panel? A
setting of \code{"on"} (the default) means yes, and a setting of \code{"off"}
means no. For details, please see \code{\link[=coord_cartesian]{coord_cartesian()}}.}
}
\value{
ggplot2 coord system
}
\description{
Customizing needs to be done
}
\seealso{
\code{ggplot2::cord_polar()}
}
