% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_radar.R
\name{coord_radar}
\alias{coord_radar}
\title{Coordinate system for ggradar()}
\usage{
coord_radar(theta = "x", start = 0, direction = 1, bg = NULL)
}
\arguments{
\item{theta}{variable to map angle to (\code{x} or \code{y})}

\item{start}{offset of starting point from 12 o'clock in radians}

\item{direction}{1, clockwise; -1, anticlockwise}

\item{bg}{raster for background image}
}
\description{
based on coord_polar(), but paths are connected linear
}
