% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_periodic.R
\name{create_percentile_functions}
\alias{create_percentile_functions}
\title{Searches a list for strings percentileXX and returns function (x, ...) quantile(x, XX, ...)}
\usage{
create_percentile_functions(fun)
}
\arguments{
\item{fun}{named list containing functions or strings}
}
\value{
named list with strings matching percentileXX replaced with functions
}
\description{
Searches a list for strings percentileXX and returns function (x, ...) quantile(x, XX, ...)
}
\keyword{internal}
