% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutfuns.R
\name{cut_daylight}
\alias{cut_daylight}
\title{Cut a POSIXct Vector into daylight (day, night)}
\usage{
cut_daylight(x, coords = c(lat = 47.36667, lon = 8.55))
}
\arguments{
\item{x}{a POSIXct Vector with a timezone}

\item{coords}{a named vector of the location in  WGS84 coordinates for the daylight factoring.
\code{\link[suncalc:getSunlightTimes]{suncalc::getSunlightTimes()}} is used to calculate sunrise, sunset times.
Default: c(lat = 47.36667, lon = 8.55) => Zuerich, Switzerland}
}
\value{
factor vector with levels day, night
}
\description{
Cut a POSIXct Vector into daylight (day, night)
}
\examples{
fn <- rOstluft.data::f("Zch_Stampfenbachstrasse_h1_2013_Jan.csv")
data <- rOstluft::read_airmo_csv(fn)

data <- dplyr::mutate(data,
  daylight_zuerich = cut_daylight(.data$starttime),
  daylight_sidney =  cut_daylight(.data$starttime, c(lat = -33.9, lon = 151.2))
)

tibble::glimpse(data)
}
