% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutfuns.R
\name{cut_season.fun}
\alias{cut_season.fun}
\title{Partial function constructor for cut_season}
\usage{
cut_season.fun(labels = NULL)
}
\arguments{
\item{labels}{a list for recoding. Names and order should be "DJF", "MAM", "JJA", "SON"}
}
\value{
Partial function of \code{\link[=cut_season]{cut_season()}} with x as sole argument
}
\description{
Partial function constructor for cut_season
}
\seealso{
\code{\link[=cut_season]{cut_season()}}
}
