% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutfuns.R
\name{cut_season}
\alias{cut_season}
\title{Cut date-time vectors into seasons}
\usage{
cut_season(x, labels = NULL)
}
\arguments{
\item{x}{a date-time vector}

\item{labels}{a list for recoding. Names and order should be "DJF", "MAM", "JJA", "SON"}
}
\value{
factor of seasons
}
\description{
Cut date-time vectors into seasons
}
\examples{
dates <- lubridate::ymd(010101) + months(0:11)

cut_season(dates)

cut_season(dates, c(DJF = "winter", MAM = "spring", JJA = "summer", SON = "autumn"))
}
