% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutfuns.R
\name{cut_seasonyear.fun}
\alias{cut_seasonyear.fun}
\title{Partial function constructor for cut_season}
\usage{
cut_seasonyear.fun(label = c("yearseason", "year"), labels = NULL)
}
\arguments{
\item{label}{choice between \code{c("yearseason", "year")}. \code{"yearseason"} will combine
the year and the output from \code{\link[=cut_season]{cut_season()}}, \code{"year"} will return only the
adjusted year.}

\item{labels}{forwarded to \code{\link[=cut_season]{cut_season()}}}
}
\value{
Partial function of \code{\link[=cut_seasonyear]{cut_seasonyear()}} with x as sole argument
}
\description{
Partial function constructor for cut_season
}
\seealso{
\code{\link[=cut_seasonyear]{cut_seasonyear()}}
}
