% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutfuns.R
\name{cut_seasonyear}
\alias{cut_seasonyear}
\title{Cut seasons, keep years together}
\usage{
cut_seasonyear(x, label = c("yearseason", "year"), labels = NULL)
}
\arguments{
\item{x}{a date-time vector}

\item{label}{choice between \code{c("yearseason", "year")}. \code{"yearseason"} will combine
the year and the output from \code{\link[=cut_season]{cut_season()}}, \code{"year"} will return only the
adjustet year.}

\item{labels}{forwarded to \code{\link[=cut_season]{cut_season()}}}
}
\value{
factor of yearseasons
}
\description{
Cut the data in year-season intervals while keeping the seasons together.
This means december will be added to the following year.

With \code{label = "year"} only the year will be adjustet.
}
\examples{
dates <- lubridate::ymd(010101) + months(0:11)

cut_seasonyear(dates)

cut_seasonyear(dates, "year")

# customize season labels
labels =  c(
  DJF = "winter", JJA = "summer",
  MAM = "spring", SON = "autumn"
)

cut_seasonyear(dates, labels = labels)
}
