% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutfuns.R
\name{cut_time_of_day}
\alias{cut_time_of_day}
\title{Cut a POSIXct Vector into time of day (HH:MM)}
\usage{
cut_time_of_day(x)
}
\arguments{
\item{x}{a datetime object.}
}
\value{
ordered factor with time of day
}
\description{
Cut a POSIXct Vector into time of day (HH:MM)
}
\examples{
fn <- rOstluft.data::f("Zch_Stampfenbachstrasse_h1_2013_Jan.csv")
data <- rOstluft::read_airmo_csv(fn)

data <- dplyr::mutate(data,
  "Etc/GMT-1" = cut_time_of_day(.data$starttime),
  utc = cut_time_of_day(lubridate::force_tz(.data$starttime, "UTC")),
  utc_converted = cut_time_of_day(lubridate::with_tz(.data$starttime, "UTC")),
  berlin = cut_time_of_day(lubridate::force_tz(.data$starttime, "Europe/Berlin"))
)

tibble::glimpse(data)
}
