% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutfuns.R
\name{cut_weekend}
\alias{cut_weekend}
\title{Cut a POSIXct Vector into weekday and weekend}
\usage{
cut_weekend(x, label = c("Wochentag", "Wochenende"))
}
\arguments{
\item{x}{a datetime object.}

\item{label}{Vector of label for weekday and weekend in order weekday, weekend}
}
\value{
ordered factor weekday > weekend
}
\description{
Cut a POSIXct Vector into weekday and weekend
}
\examples{
fn <- rOstluft.data::f("Zch_Stampfenbachstrasse_d1_2013_Jan.csv")
data <- rOstluft::read_airmo_csv(fn)

data <- dplyr::mutate(data,
  weekend_ger = cut_weekend(.data$starttime),
  weekend_eng =  cut_weekend(.data$starttime, c("weekday", "weekend"))
)

tibble::glimpse(data)
}
