% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_cartpolar.R
\name{element_render_circle}
\alias{element_render_circle}
\title{renders a circle as element_line}
\usage{
element_render_circle(theme, element, x, y, r, ..., name = NULL,
  nseg = 360)
}
\arguments{
\item{theme}{A \code{theme} object}

\item{element}{The name of a theme element (see \code{\link[ggplot2:calc_element]{ggplot2::calc_element()}}).}

\item{x}{a numeric vector specifying x center locations of circles}

\item{y}{a numeric vector specifying y center locations of circles}

\item{r}{a numeric vector specifying radii of circles}

\item{...}{passed onto \code{\link[ggplot2:element_grob]{ggplot2::element_grob()}}}

\item{name}{of generated grid object}

\item{nseg}{number of line segments to draw the circle}
}
\value{
\code{\link[ggplot2:zeroGrob]{ggplot2::zeroGrob()}} or \code{\link[grid:polylineGrob]{grid::polylineGrob()}}
}
\description{
renders a circle as element_line
}
