% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_sprintf}
\alias{format_sprintf}
\title{Wrapper for sprintf string formatting}
\usage{
format_sprintf(fmt)
}
\arguments{
\item{fmt}{sprintf formatting string. see \code{sprintf()}}
}
\value{
function formatting x with sprintf
}
\description{
Can be used to provide format functions for labels in ggpltos
}
\examples{
# add one leading zeroes for 1 digits numbers
add_leading_zeroes <- format_sprintf("\%02d")
add_leading_zeroes(1:10)
}
