% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ggmap.R
\name{get_ggmap}
\alias{get_ggmap}
\title{Wrapper to load a ggmap}
\usage{
get_ggmap(coord, googlekey = NA, zoom = 12, mapstyle = c(feature =
  "all", element = "labels", visibility = "off"), maptype = "terrain",
  mapcolor = "bw", ...)
}
\arguments{
\item{coord}{a data.frame or tibble with coordinates (latlon) defining the map extensions () in the form of coord = tibble(lon = ..., lat = ..., type = c("center", "lb", "rt", ...));
the reason for this structure is, that ggmap::get_googlemap requires a different way (center + zoom) to specify the map frame than ggmap::get_stamenmap (bounding box). If needed, coord is
converted to bounding box by rOstluft.plot::bbox() which takes type == "lb" as bottom & left coordinate, and "rt" as top & right, respectively.}

\item{googlekey}{yourt google api key as a character string (see ?ggmap::get_googlemap; then, a google map is downloaded) or, if NA, a map via  ggmap::get_stamenmap is returned}
}
\value{
ggplot / ggmap object
}
\description{
Wrapper to load a ggmap
}
