% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{get_google_map}
\alias{get_google_map}
\title{Wrapper around ggmap::get_googlemap}
\usage{
get_google_map(bbox, width = 640, zoom = "auto", scale = 2,
  language = "de-CH", color = "bw", style = c(feature = "all",
  element = "labels", visibility = "off"), ...)
}
\arguments{
\item{bbox}{list with item \code{center = (lon, lat)} as WSG84 coordinates. Additional \code{width} in meters if \code{zoom = "auto"}.
Use \code{\link[=bbox_lv95]{bbox_lv95()}} to create a compatible bounding box list}

\item{width}{of resulting map in pixels. Is divided trough scale to calculate size.}

\item{zoom}{of tile map as Integer or "auto"}

\item{scale}{affects the size of labels. \code{\link[ggmap:get_googlemap]{ggmap::get_googlemap()}} multiplies the size with this factor.}

\item{language}{character string providing language of map labels (for themes with them) in the format "en-EN".
not all languages are supported; for those which aren't the default language is used.}

\item{color}{color or black-and-white. Changed default to "bw"}

\item{style}{character string or named vector to style the map. This is a powerful and complex specification. See
\href{https://developers.google.com/maps/documentation/maps-static/styling#style-syntax}{Styled Maps} in the documentation of
the google maps platform}

\item{...}{forwarded to \code{\link[ggmap:get_googlemap]{ggmap::get_googlemap()}}}
}
\value{
a ggmap object (a classed raster object with a bounding box attribute)
}
\description{
The functions \code{\link[=get_googlemap]{get_googlemap()}} and \code{\link[=get_stamen_map]{get_stamen_map()}} tries to harmonize the calls to \code{\link[=ggmap]{ggmap()}}. The objective is
to get interchangeable functions with sensible defaults. For example automatic calculation of the zoom. The function
\code{\link[=bbox_lv95]{bbox_lv95()}} generates a bbox object compatible with both functions. Before using this function you need to register
your google API key with \code{\link[ggmap:register_google]{ggmap::register_google()}}. For details consult the documentation of the
\href{https://developers.google.com/maps/documentation/maps-static/intro}{Maps Static API} and \code{\link[ggmap:get_googlemap]{ggmap::get_googlemap()}}.
}
\examples{
library(ggmap)
ggmap::register_google(Sys.getenv("GGMAP_GOOGLE_API_KEY"))
bb <- bbox_lv95(2683141, 1249040, 500) # site Zch_Stamfenbachstrasse
get_google_map(bb) \%>\%  ggmap::ggmap()
}
