% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{get_stadia_map}
\alias{get_stadia_map}
\title{Wrapper around ggmap::get_stadiamap()}
\usage{
get_stadia_map(
  bbox,
  maptype = "stamen_terrain",
  width = 640,
  zoom = "auto",
  color = "bw",
  ...
)
}
\arguments{
\item{bbox}{list with items \code{left}, \code{bottom}, \code{right} and \code{top} as WSG84 coordinates. Additional \code{width}
in meters if \code{zoom = "auto"}. Use \code{\link[=bbox_lv95]{bbox_lv95()}} to create a compatible bounding box list}

\item{maptype}{stamen_terrain, stamen_toner, stamen_toner_lite, stamen_watercolor, stamen_terrain_background, stamen_toner_background, stamen_terrain_lines, stamen_terrain_labels, stamen_toner_lines, stamen_toner_labels.}

\item{width}{of resulting map in pixels.}

\item{zoom}{zoomlevel from 0 to 18 or "auto"}

\item{color}{color or black-and-white. Changed default to "bw"}

\item{...}{forwarded to \code{\link[ggmap:get_stadiamap]{ggmap::get_stadiamap()}}}
}
\value{
a ggmap object (a classed raster object with a bounding box attribute)
}
\description{
The functions \code{\link[=get_googlemap]{get_googlemap()}} and \code{\link[=get_stadia_map]{get_stadia_map()}} tries to harmonize the calls to \code{\link[=ggmap]{ggmap()}}. The objective is
to get interchangeable functions with sensible defaults. For example automatic calculation of the zoom. The function
\code{\link[=bbox_lv95]{bbox_lv95()}} generates a bbox object compatible with both functions.
}
\section{API-Key}{

To obtain an API key and enable services, go to \url{https://client.stadiamaps.com/signup/}. It is completely free for non-commercial and evaluation use (a license is for commercial use; see https://stadiamaps.com/pricing for pricing), and no credit card is required to sign up.
For details see \code{\link[ggmap:register_stadiamaps]{ggmap::register_stadiamaps()}}.
}

\section{Attribution}{

Details on attribution can be found at \url{https://stadiamaps.com/attribution}.
}

\examples{
bb <- bbox_lv95(2683141, 1249040, 500) # site Zch_Stamfenbachstrasse
get_stadia_map(bb) |>
  ggmap::ggmap()
}
