% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ggmap_stamen.R
\name{get_stamen_map}
\alias{get_stamen_map}
\title{Wrapper around ggmap::get_stamen_map}
\usage{
get_stamen_map(bbox, width = 640, zoom = "auto", color = "bw", ...)
}
\arguments{
\item{bbox}{list with items \code{left}, \code{bottom}, \code{right} and \code{top} as WSG84 coordinates. Additional \code{width}
in meters if \code{zoom = "auto"}. Use \code{\link[=bbox_lv95]{bbox_lv95()}} to create a compatible bounding box list}

\item{width}{of resulting map in pixels.}

\item{zoom}{zoomlevel from 0 to 18 or "auto"}

\item{color}{color or black-and-white. Changed default to "bw"}

\item{...}{forwarded to \code{\link[ggmap:get_googlemap]{ggmap::get_googlemap()}}}
}
\value{
a ggmap object (a classed raster object with a bounding box attribute)
}
\description{
Wrapper around ggmap::get_stamen_map
}
\examples{
bb <- bbox_lv95(2683141, 1249040, 500) # site Zch_Stamfenbachstrasse
get_stamen_map(bb) \%>\% ggmap::ggmap()
}
