% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggperiodic.R
\name{ggdiurnal_weekend}
\alias{ggdiurnal_weekend}
\title{ggplot2-wrapper to summarise and plot data (of \code{\link[rOstluft:format_rolf]{rOstluft::format_rolf()}} format) for diurnal time course-stats
including weekend-effect}
\usage{
ggdiurnal_weekend(data, y = "value", nmin = 3, fun = list(middle =
  "mean", middle2 = "median", ymin = "percentile05", lower =
  "percentile25", upper = "percentile75", ymax = "percentile95"),
  ribbon_color = "gray60", middle_color = "gray30",
  expand_xscale = 0)
}
\arguments{
\item{data}{a data.frame or tibble containing the data}

\item{y}{a character string specifying the taget variable to be summarised, usually "value".
outcomes of \code{\link[rOstluft:cut_timeseries_periodic]{rOstluft::cut_timeseries_periodic()}}.}

\item{fun}{function or list of functions for summary, can be named (then the outut stat is named after the function's name);
Strings matching the regular expression \code{/^percentile([0-9]){1,2}$/i} are converted into the respective function.
"percentile95" => \code{function(x, ...) quantile(x, 95 / 100, ...)}}

\item{ribbon_color}{character string for specifying the ribbon color.}

\item{middle_color}{character string for specifying the middle's color.}

\item{expand_xscale}{multiplicative range expansion factor}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
ggplot2-wrapper to summarise and plot data (of \code{\link[rOstluft:format_rolf]{rOstluft::format_rolf()}} format) for diurnal time course-stats
including weekend-effect
}
\examples{
fn <- rOstluft.data::f("Zch_Stampfenbachstrasse_2010-2014.csv")

data <-
  rOstluft::read_airmo_csv(fn) \%>\%
  dplyr::filter(parameter \%in\% c("NOx", "O3", "PM10")) \%>\%
  rOstluft::resample(new_interval = "h1")

ggdiurnal_weekend(data)

# y scale: no expand on zero
ggdiurnal_weekend(data) +
  scale_y_continuous(limits = c(0, NA), expand = expand_scale(mult = c(0, 0.05))) +
  theme(panel.spacing.y = unit(12, "pt"))
}
