% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggperiodic.R
\name{ggdiurnal_weekend_diff}
\alias{ggdiurnal_weekend_diff}
\title{ggplot2-wrapper to summarise and plot data (of \code{\link[rOstluft:format_rolf]{rOstluft::format_rolf()}} format) for one diurnal time course-stat
plotted as difference between weekdays - weekend}
\usage{
ggdiurnal_weekend_diff(data, y = "value", nmin = 3, fun = list(middle
  = "mean"), expand_xscale = 0.025, ...)
}
\arguments{
\item{data}{a data.frame or tibble containing the data}

\item{y}{a character string specifying the taget variable to be summarised, usually "value".
outcomes of \code{\link[rOstluft:cut_timeseries_periodic]{rOstluft::cut_timeseries_periodic()}}.}

\item{fun}{function or list of functions for summary, can be named (then the outut stat is named after the function's name);
Strings matching the regular expression \code{/^percentile([0-9]){1,2}$/i} are converted into the respective function.
"percentile95" => \code{function(x, ...) quantile(x, 95 / 100, ...)}}

\item{expand_xscale}{multiplicative range expansion factor}

\item{...}{further arguments passed on to \code{\link[ggplot2:geom_line]{ggplot2::geom_line()}}.}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
ggplot2-wrapper to summarise and plot data (of \code{\link[rOstluft:format_rolf]{rOstluft::format_rolf()}} format) for one diurnal time course-stat
plotted as difference between weekdays - weekend
}
\examples{
fn <- rOstluft.data::f("Zch_Stampfenbachstrasse_2010-2014.csv")

data <-
  rOstluft::read_airmo_csv(fn) \%>\%
  dplyr::filter(parameter \%in\% c("NOx", "O3", "PM10")) \%>\%
  rOstluft::resample(new_interval = "h1")

ggdiurnal_weekend_diff(data)
}
