% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggperiodic.R
\name{ggweekend}
\alias{ggweekend}
\title{ggplot2-wrapper to summarise and plot data (of \code{\link[rOstluft:format_rolf]{rOstluft::format_rolf()}} format) as boxplots
comparing weekdays to weekends}
\usage{
ggweekend(data, y = "value", nmin = 3, fun = list(middle = "mean",
  middle2 = "median", ymin = "percentile05", lower = "percentile25", upper
  = "percentile75", ymax = "percentile95"), ...)
}
\arguments{
\item{data}{a data.frame or tibble containing the data}

\item{y}{a character string specifying the taget variable to be summarised, usually "value".
outcomes of \code{\link[rOstluft:cut_timeseries_periodic]{rOstluft::cut_timeseries_periodic()}}.}

\item{fun}{function or list of functions for summary, can be named (then the outut stat is named after the function's name);
Strings matching the regular expression \code{/^percentile([0-9]){1,2}$/i} are converted into the respective function.
"percentile95" => \code{function(x, ...) quantile(x, 95 / 100, ...)}}

\item{...}{further arguments passed on to \code{\link[ggplot2:geom_boxplot]{ggplot2::geom_boxplot()}}.}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
ggplot2-wrapper to summarise and plot data (of \code{\link[rOstluft:format_rolf]{rOstluft::format_rolf()}} format) as boxplots
comparing weekdays to weekends
}
\examples{
fn <- rOstluft.data::f("Zch_Stampfenbachstrasse_2010-2014.csv")

data <-
  rOstluft::read_airmo_csv(fn) \%>\%
  dplyr::filter(parameter \%in\% c("NOx", "O3", "PM10")) \%>\%
  rOstluft::resample(new_interval = "h1")

ggweekend(data)

# change y axis title using bquote
ggweekend(data) +
  ylab(bquote("Konzentration " ~ 0[3] ~ ", " ~ PM[10] ~ "(" ~ mu ~ "g " ~ m^{-3} ~ "), NOx (ppb)" ))

# change fill scale
ggweekend(data) +
  scale_fill_brewer(type = "qual", palette = "Pastel1")
}
