% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggwindrose.R
\name{ggwindrose2}
\alias{ggwindrose2}
\title{ggplot wrapper to create a windrose with calm support (polar wind-bin frequency plot)}
\usage{
ggwindrose2(
  data,
  ws,
  wd,
  wd_binwidth = 45,
  ws_binwidth = 1,
  ws_max = NA,
  groupings = grp(),
  nmin = 1,
  fill_scale = scale_fill_viridis_d(),
  reverse = TRUE,
  bg = NULL,
  calm = NA,
  calm.prefix = "Calm\\n",
  calm.accuracy = 1,
  calm.color = "black",
  calm.size = 3.88,
  yexpand = 0.2,
  ...
)
}
\arguments{
\item{data}{tibble containing wind speed, wind direction and/or air pollutant concentration data}

\item{ws}{symbol giving the wind velocity column name (wind velocity preferably in m/s)}

\item{wd}{symbol giving the wind direction column name  in degrees}

\item{wd_binwidth}{numeric, binwidth for wind direction in °, wd_binwidth should fullfill:
\code{(360 / wd_binwidth) \%in\% c(4, 8, 16, 32)}}

\item{ws_binwidth}{numeric, binwidth for wind speed}

\item{ws_max}{numeric, can be NA, wind speed is squished at this value}

\item{groupings}{additional groupings. Use helper \code{\link[=grp]{grp()}} to create. \strong{Necessary} for some facets!}

\item{nmin}{numeric, minimum number of values for fun, if n < nmin: NA is returned}

\item{fill_scale}{ggplot2 discrete fill scale, e.g. \code{\link[ggplot2:scale_gradient]{ggplot2::scale_fill_gradientn()}}}

\item{reverse}{TRUE/FALSE, should wind speed bin factors be sorted descending (TRUE)
or ascending (FALSE). Usually for wind roses a descending order (higher wind speed on
the outside) is used.}

\item{bg}{raster map, e.g. ggmap object as plot background}

\item{calm}{if specified as numeric value all windspeed below this value will be treated
as calm and the proportion will be displayed as text in the center. The color, size of
the text can be supplied with \code{calm.size} and \code{calm.color}. A prefix for the text
can be defined with \code{calm.prefix} and the accuracy of the number with \code{calm.accuracy}}

\item{calm.prefix}{prefix for the calm text as string. Supplied to \code{\link[scales:label_percent]{scales::label_percent()}} as
argument.}

\item{calm.accuracy}{A number to round to. Use (e.g.) 0.01 to show 2 decimal places of
precision. If NULL uses a heuristic that should ensure breaks have the minimum number
of digits needed to show the difference between adjacent values.}

\item{calm.color}{text color}

\item{calm.size}{text size}

\item{yexpand}{size of the empty calm circle in the center as percentage of the y scale}

\item{...}{Other arguments passed on to \code{\link[ggplot2:geom_bar]{ggplot2::geom_bar()}}. Used
to set an aesthetic to a fixed value. Defaults are \verb{color = "white", width = 1, size = 0.25}}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
ggplot wrapper to create a windrose with calm support (polar wind-bin frequency plot)
}
\examples{
library(ggplot2)

fn <- rOstluft.data::f("Zch_Stampfenbachstrasse_2010-2014.csv")

data <-
  rOstluft::read_airmo_csv(fn) \%>\%
  rOstluft::rolf_to_openair()

ggwindrose2(data, ws, wd, calm = 0.3)

# squish ws
ggwindrose2(data, ws, wd, ws_max = 5, calm = 0.3)

#' # a map as background
bb <- bbox_lv95(2683141, 1249040, 500)
bg <- get_stamen_map(bb)
ggwindrose2(data, ws, wd, ws_max = 5, calm = 0.5, alpha = 0.8, bg = bg) +
  theme(
    panel.grid.major = element_line(linetype = 2, color = "black", size = 0.5)
   )
}
