% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggyearday.R
\name{ggyearday}
\alias{ggyearday}
\title{ggplot heatmap for diurnal-yearly time course}
\usage{
ggyearday(data, time, z, xbreaks = "1 month", xlabels = "\%b",
  ybreaks = seq(3, 21, 3), ylabels = format_sprintf("\%02d:00"),
  fill_scale = scale_fill_viridis_c(direction = -1, na.value = NA, option
  = "A"), ...)
}
\arguments{
\item{data}{a data.frame or tibble with input data (containing a POSIXct variable as time parameter).}

\item{time}{character string giving time variable name.}

\item{z}{character string giving z value variable name.}

\item{...}{Other arguments passed on to geom_raster().}

\item{etc}{...}
}
\value{
ggplot
}
\description{
Create a heatmap with date on x-axis and time of day on y-axis; z values as colour scale.
}
\examples{
require(rOstluft)
require(rOstluft.data)
require(rOstluft.plot)
require(ggplot2)
require(dplyr)

df <-
  rOstluft::read_airmo_csv(system.file("extdata", "Zch_Stampfenbachstrasse_2010-2014.csv", package = "rOstluft.data", mustWork = TRUE)) \%>\%
  rOstluft::rolf_to_openair() \%>\%
  mutate(year = lubridate::year(date))

ggyearday(df, time = "date", z = "O3") +
  facet_wrap(year~., scales = "free_x", ncol = 1)

# data with outliers / extreme values => not very informative...
ggyearday(df, time = "date", z = "PM10") +
  facet_wrap(year~., scales = "free_x", ncol = 1)

# ...use a custom scale and squish the outliers / extreme values
fill_scale <- scale_fill_viridis_squished(breaks=c(0, 25, 50, 75), limits = c(0, 75),
                                          direction = -1, na.value = NA, option = "A")
ggyearday(df, time = "date", z = "PM10", fill_scale = fill_scale) +
  facet_wrap(year~., scales = "free_x", ncol = 1)


}
