% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grp}
\alias{grp}
\title{Pass grouping definitions into functions}
\usage{
grp(...)
}
\arguments{
\item{...}{Variables to group by. These arguments are automatically
\link[rlang:defusing-advanced]{quoted} and later \link[rlang:eval_tidy]{evaluated} in the
context of the data frame. They support \link[rlang:topic-inject]{unquoting}.}
}
\value{
named list containing quosures or symbols
}
\description{
Inspired by \code{\link[dplyr:vars]{dplyr::vars()}}, but converting strings to symbols and auto names all arguments
}
\examples{
fn <- rOstluft.data::f("Zch_Stampfenbachstrasse_d1_2017.csv")
data <- rOstluft::read_airmo_csv(fn)

# adding group_nest to quickly glance over the groups
groupby <- function(df, group = grp()) {
  dplyr::group_by(df, !!!group) \%>\%
    dplyr::group_nest()
}

# no grouping -> everything will be nested
groupby(data)

# use a symbol, string or an expression
groupby(data, grp(site, "unit", lubridate::year(starttime)))

# autonaming works fine with strings and symbols, but for expressions
# it probably a good idea to provide a name:
groupby(data, grp(site, year = lubridate::year(starttime)))
}
