% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_factor_funs.R
\name{midpoints}
\alias{midpoints}
\title{returns class midpoint-values for '[x-y)'-style factor levels}
\usage{
midpoints(x)
}
\description{
returns class midpoint-values for '[x-y)'-style factor levels
}
