% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcalendar.R
\name{pad_to_year}
\alias{pad_to_year}
\title{Pads data to complete year
additional argument to fill some columns like:
https://github.com/davidcarslaw/openair/blob/592a076491791505b0ffcde099d84e7973f8b1c8/R/utilities.R#L78}
\usage{
pad_to_year(data, date_col, interval)
}
\arguments{
\item{data}{input data}

\item{date_col}{column containing data}

\item{interval}{interval between two dates}
}
\value{
padded data
}
\description{
Pads data to complete year
additional argument to fill some columns like:
https://github.com/davidcarslaw/openair/blob/592a076491791505b0ffcde099d84e7973f8b1c8/R/utilities.R#L78
}
