% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad.R
\name{pad_to_year_fill}
\alias{pad_to_year_fill}
\title{Pads data to complete year and fill up specific columns}
\usage{
pad_to_year_fill(data, date_col, interval, ...)
}
\arguments{
\item{data}{input data}

\item{date_col}{column containing date information}

\item{interval}{interval between two dates}

\item{...}{columns to fill up}
}
\value{
padded data
}
\description{
Pads data to complete year and fill up specific columns
}
