% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{plt_cal}
\alias{plt_cal}
\alias{cal_month_border}
\alias{cal_label}
\title{calendar plot}
\usage{
plt_cal(data, x = "date", z = "O3_max_h1", size = 0.1,
  color = "white", ..., locale = Sys.getlocale("LC_TIME"))

cal_month_border(size = 1, lineend = "square", linejoin = "bevel",
  color = "grey5", ...)

cal_label(mapping = NULL, stat = "identity", data = NULL,
  geom = "text", position = "identity", show.legend = FALSE,
  inherit.aes = TRUE, na.rm = TRUE, size = 2, color = "white", ...)
}
\arguments{
\item{data}{input data}

\item{x}{date column as Date, POSIXct or Character}

\item{z}{value columns}

\item{size}{space between tiles}

\item{color}{color of space}

\item{...}{more options for \code{ggplot2::geom_tile()}}

\item{locale}{locale string for \code{lubridate::month()} and \code{lubridate::wday()}. see \code{Sys.getlocale()}}

\item{lineend}{Line end style (round, butt, square).}

\item{linejoin}{Line join style (round, mitre, bevel).}

\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{geom}{geom used for layer. "label" could be an alternative}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}
}
\value{
ggplot2 object

ggplot2 layer

ggplot2 layer
}
\description{
TODO: example

Just a geom_text with some defaults
}
