% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcalendar.R
\name{plt_calendar}
\alias{plt_calendar}
\title{Plotting data on a calendar}
\usage{
plt_calendar(data, x = "starttime", z = "value", tile_opt = list(size
  = 0.1, color = "white"), scale_fill = scale_fill_viridis_c(),
  label = monthday, label_opt = list(size = 2, color = "white"),
  border = TRUE, border_opt = list(size = 1, lineend = "square",
  linejoin = "bevel", color = "grey5"), marker = NULL,
  marker_opt = list(size = 2, position = position_nudge(y = 0.3), color =
  "white"), locale = Sys.getlocale("LC_TIME"))
}
\arguments{
\item{data}{input data}

\item{x}{date column as Date, POSIXct or Character}

\item{z}{value columns}

\item{tile_opt}{geom_tile params. see \code{ggplot2::geom_tile()} and \code{ggplot2::layer()}}

\item{scale_fill}{scale for tile fill. Default is for continous data. For discrete data one must be provided}

\item{label}{expression for label or NULL}

\item{label_opt}{geom_text params. see \code{ggplot2::geom_text()} and \code{ggplot2::layer()}}

\item{border}{draw a border for every month. TRUE or FALSE}

\item{border_opt}{geom_segement params. see \code{ggplot2::geom_segment()} and \code{ggplot2::layer()}}

\item{marker}{expression for placing marker (\code{dplyr::filter()}) or NULL}

\item{marker_opt}{geom_point params. see \code{ggplot2::geom_segment()} and \code{ggplot2::layer()}}

\item{locale}{locale string. see \code{Sys.getlocale()}}
}
\value{
ggplot2 object
}
\description{
Calculated Columns:
\itemize{
\item monthday: day of month
\item month: factor value = 1-12, level = localized abbreviated name
\item year
\item weekday: factor value = 1-7, level = localized name
\item week
\item x: year x 100 + week
}

We need a continous x for placing breaks for every month on every facet
}
