% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_factor_funs.R
\name{recode_last_class_label}
\alias{recode_last_class_label}
\title{recode highest \code{[x-y)}-style factor level to yield '>y'}
\usage{
recode_last_class_label(factor_var)
}
\arguments{
\item{factor_var}{factor from cutting function to recode}
}
\description{
recode highest \code{[x-y)}-style factor level to yield '>y'
}
