% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_cartpolar.R, R/coord_polar2.R,
%   R/geom_bar_wind.R, R/geom_polar_raster.R, R/ggcalendar.R,
%   R/rOstluft.plot-package.R, R/stat_filter.R, R/stat_summary_wind.R,
%   R/stat_summary_wind_2d.R
\docType{data}
\name{CoordCartPolar}
\alias{CoordCartPolar}
\alias{CoordPolar2}
\alias{GeomBarWind}
\alias{GeomPolarRaster}
\alias{CalMonthBorder}
\alias{rOstluft-ggproto}
\alias{StatFilter}
\alias{StatSummaryWind}
\alias{StatSummaryWind2d}
\title{All ggproto Objects}
\format{An object of class \code{CoordCartPolar} (inherits from \code{CoordCartesian}, \code{Coord}, \code{ggproto}, \code{gg}) of length 11.}
\usage{
CoordCartPolar

CoordPolar2

GeomBarWind

GeomPolarRaster

CalMonthBorder

StatFilter

StatSummaryWind

StatSummaryWind2d
}
\description{
All ggproto Objects
}
\seealso{
\itemize{
\item \link[ggplot2:ggproto]{ggplot2::ggproto}
\item \href{https://ggplot2.tidyverse.org/reference/ggplot2-ggproto.html}{Base ggproto classes for ggplot2}
}
}
\keyword{datasets}
\keyword{internal}
