% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rOstluft.plot-package.R
\docType{package}
\name{rOstluft.plot-package}
\alias{rOstluft.plot}
\alias{rOstluft.plot-package}
\title{rOstluft.plot: Ploting air quality data}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

This package provides functions to quickly display data as diagrams.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Ostluft/rOstluft.plot}
  \item Report bugs at \url{https://github.com/Ostluft/rOstluft.plot/issues}
}

}
\author{
\strong{Maintainer}: Thomas von Allmen \email{thomas.vonallmen@zuerich.ch}

Authors:
\itemize{
  \item Jörg Sintermann
}

Other contributors:
\itemize{
  \item Ostluft [copyright holder]
}

}
\keyword{internal}
