% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viridis_squished.R
\name{squished_labels}
\alias{squished_labels}
\title{generates labels for squished scales}
\usage{
squished_labels(labels)
}
\arguments{
\item{labels}{One of:
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels computed by the
transformation object
\item A character vector giving labels (must be same length as \code{breaks})
\item A function that takes the breaks as input and returns labels
as output
Additional if not NULL the last element will be formated with a leading ">"
}}
}
\value{
function, labels for scales
}
\description{
generates labels for squished scales
}
