% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_summary_wind.R
\name{stat_summary_wind}
\alias{stat_summary_wind}
\title{Summarise z values over binned wind data.}
\usage{
stat_summary_wind(data = NULL, mapping = NULL, geom = "polygon",
  position = "identity", fun = "mean", fun.args = list(),
  show.legend = NA, inherit.aes = TRUE, nmin = 1,
  wd_cutfun = function(wd) wd_classes(wd, wd_binwidth = 45),
  wd_offset = 0, ws_cutfun = function(ws) ws_classes(ws, ws_binwidth =
  1, ws_max = NA), groups = NULL, ...)
}
\arguments{
\item{data}{The data to be displayed in this layer.
#' requires input data including at least three columns carrying information regarding:
\itemize{
\item wind direction
\item wind velocity
\item z
}}

\item{mapping}{Set of aesthetic mappings created by aes() or aes_().
If specified and inherit.aes = TRUE (the default), it is combined with
the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping..}

\item{geom}{The geometric object to use display the data.}

\item{fun}{function or list of functions for summary.}

\item{fun.args}{A list of extra arguments to pass to fun.}

\item{nmin}{Minimum number of values for fun, if n < nmin: NA is returned}

\item{groups}{can be NULL, c("u", "v"), ...}

\item{...}{Other arguments passed on to layer(params = list(...)).}

\item{ws_max}{Maximum wind velocity for binning: above ws_max, z is set NA}

\item{bins}{number of bins over the range of values if !groups %in% c("u", "v")}

\item{smooth}{TRUE/FALSE, applies if groups = c("u", "v"); should smoothing of summary results should be performed
using gam_surface()?}

\item{k}{numeric, applies if smooth = TRUE; degree of smoothing in gam_surface()}

\item{extrapolate}{TRUE/FALSE, applies if smooth = TRUE; gem_smooth() returns extrapolated values for u, v coordinates that have NA for summarised z
if extrapolate = TRUE, those values are returned (to a certain degree depending on the value of dist)}

\item{dist}{numeric, fraction of 1, applies if smooth = TRUE and extrapolate = TRUE; maximum distance to coordinate-pair at which the result of
gem_smooth(z) should be returned}
}
\value{
ggplot2 layer

Aesthetics
\itemize{
\item wd: wind direction in degrees
\item ws: wind velocity
\item z: z values to be summarised
}

Computed variables
\itemize{
\item If groups = c("u", "v"): a tibble is returned, binned over u and v, with variables:
}
\itemize{
\item wd: wind direction corresponding to midpoint value of u and v
\item ws: wind velocity corresponding to midpoint value of u and v
\item u: bins over u (from input wd and ws)
\item v: bins over v (from input wd and ws)
\item z: result from fun(z, ...)
}
\itemize{
\item groups can be strings for other variables in data; then fun is applied over those;
a tibble including groups and summarised z is returned
}
}
\description{
Binning is done by StatWind, so input data to stat_summary_wind() should be original unbinned data.
Depending on the groups argument, binning is either done 2-dimensional over cartesian u and v wind vectors
(calculated from input data; then, stat_summary_wind() yields results similar to openair::polarplot())
or,
1-dimensional over wind direction or wind velocity bins, respectively.
}
