% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_summary_wind.R
\docType{data}
\name{StatWind}
\alias{StatWind}
\title{ggproto for stat_bin_wind()}
\format{An object of class \code{StatWind} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.}
\usage{
StatWind
}
\description{
ggproto for stat_bin_wind()
}
\keyword{datasets}
