% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_summary_wind_2d.R
\docType{data}
\name{StatWind2d}
\alias{StatWind2d}
\title{ggproto for stat_bin_wind_2d()}
\format{An object of class \code{StatWind2d} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.}
\usage{
StatWind2d
}
\description{
ggproto for stat_bin_wind_2d()
}
\keyword{datasets}
