% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uv2wd}
\alias{uv2wd}
\title{calculate wind direction from u, v wind components}
\usage{
uv2wd(u, v)
}
\arguments{
\item{u}{a vector of u components}

\item{v}{a vector of v components}
}
\value{
a vector of wind direction in °
}
\description{
copied from https://github.com/environmentalinformatics-marburg/Rsenal/blob/master/R/uv2wdws.R
}
