% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutfuns.R
\name{ws_classes}
\alias{ws_classes}
\title{wrapper to cut wind velocity (or others) into factor classes}
\usage{
ws_classes(ws, binwidth = 1, ws_max = NA, ...)
}
\description{
wrapper to cut wind velocity (or others) into factor classes
}
