% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutfuns.R
\name{y_classes}
\alias{y_classes}
\title{wrapper to cut y data into factor classes}
\usage{
y_classes(y, binwidth, ymax = NA, boundary = 0, ...)
}
\arguments{
\item{y}{a numeric vector}

\item{binwidth}{for \code{\link[ggplot2:cut_width]{ggplot2::cut_width()}}}

\item{ymax}{cut off at this maximum}

\item{boundary}{for \code{\link[ggplot2:cut_width]{ggplot2::cut_width()}}}

\item{...}{passed to \code{\link[ggplot2:cut_width]{ggplot2::cut_width()}}}
}
\description{
Wraps \code{\link[ggplot2:cut_width]{ggplot2::cut_width()}} function
}
