% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_wind_density.R
\name{zzz_plt_wind_density}
\alias{zzz_plt_wind_density}
\title{Plotting wind density}
\usage{
zzz_plt_wind_density(data)
}
\arguments{
\item{data}{in rolf format containing WD and WVv in h1 Interval}
}
\value{
ggplot object
}
\description{
Quick and dirty demonstration of concept. Needs some work
}
\examples{
h1 <- system.file("extdata", "Zch_Stampfenbachstrasse_h1_2013_Jan.csv",
                  package = "rOstluft.data", mustWork = TRUE)

data <- rOstluft::read_airmo_csv(h1)
plt_wind_density(data)
}
\keyword{internal}
