% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_wind_rose.R
\name{zzz_plt_wind_rose}
\alias{zzz_plt_wind_rose}
\title{Plotting a windrose on a stamen terrain plot}
\usage{
zzz_plt_wind_rose(data, meta)
}
\arguments{
\item{data}{in rolf format with WD and WVv}

\item{meta}{tibble with one row containing x and y coordinates of the site}
}
\value{
png image
}
\description{
Plotting a windrose on a stamen terrain plot
}
\keyword{internal}
