% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_cartpolar.R
\name{coord_cartpolar}
\alias{coord_cartpolar}
\title{cartesian coordinate system with polar grid lines}
\usage{
coord_cartpolar(limit = NULL, expand = TRUE, clip = "on",
  bg = NULL, scale_wd = NULL)
}
\arguments{
\item{limit}{limit for coordsystem (xlim(-limit, limit), ylim(-limit, limit))}

\item{expand}{expand limits (probably not used at the moment)}

\item{clip}{clip panel (not sure if used)}

\item{bg}{raster for background image}

\item{scale_wd}{pass breaks angle lines (not used at the moment)}
}
\description{
WIP:
\itemize{
\item render_axis_h / render_axis_v probably needs an overwrite
\item labels for polar grid lines
\item wd scale / breaks but how exactly?
}
}
\examples{
library(ggplot2)

df <- tibble::tribble(
  ~wd, ~ws, ~facet,
  0, 4, 1,
  90, 4, 1,
  180, 4, 1,
  270, 4, 1,
  45, 8, 2,
  135, 8, 2,
  225, 8, 2,
  315, 8, 2
)

df <- dplyr::mutate(df,
  u = .data$ws * sin(2 * pi * .data$wd / 360),
  v = .data$ws * cos(2 * pi * .data$wd / 360)
)

p <- ggplot(df, aes(x=u, y=v)) + geom_point() + coord_cartpolar()
p


# theming for panel.grid.major.x and panel.grid.minor.x is applied to polar grid lines
p + theme(
      "panel.grid.minor.x" = element_line(color = "blue"),
      "panel.grid.major.x" = element_line(color = "red", size = 2)
    )

# x breaks are polar grid radiuses:
p + scale_x_continuous(breaks = c(1, 1.5, 3, 6)) +
  theme("panel.grid.minor.x" = element_blank())


# background map
bbox <- tibble::tibble(x = c(2683141 - 500, 2683141 + 500), y = c(1249040 - 500, 1249040 + 500))
bbox <- rOstluft::transform_projection(bbox, coord = c("x", "y"),
                                       initCRS = sp::CRS("+init=epsg:2056"),
                                       outCRS = sp::CRS("+init=epsg:4326"))

bbox <- c(left = bbox$x[1], right = bbox$x[2], bottom = bbox$y[1], top = bbox$y[2])

raster_map <- ggmap::get_stamenmap(bbox, zoom = 16, maptype = "terrain",
                                   source = "stamen", color = "bw")

pbg <- ggplot(df, aes(x=u, y=v)) + geom_point() + coord_cartpolar(bg = raster_map)
pbg


# faceting
pbg + facet_wrap(~facet)


# if plotting a raster layer, use annotation_raster, panel.ontop and panel.background
df <- expand.grid(u = -5:5, v = -5:5)
df$z <- runif(nrow(df))
ggplot(df, aes(x=u, y=v, fill=z)) + coord_cartpolar(expand = FALSE) +
  annotation_raster(raster_map, -Inf, Inf, -Inf, Inf) +
  geom_raster(alpha = 0.5) +
  theme(
    "panel.ontop" = TRUE,
    "panel.background" = element_blank()
  )

}
