% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_radar.R
\name{coord_radar}
\alias{coord_radar}
\title{Coordinate system for ggradar()}
\usage{
coord_radar(theta = "x", start = 0, direction = 1, bg = NULL)
}
\description{
based on coord_polar(), but paths are connected linear
}
