% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_radar.R
\name{coord_radar}
\alias{coord_radar}
\title{Polar coordinate system without munching}
\usage{
coord_radar(theta = "x", start = 0, direction = 1, bg = NULL)
}
\arguments{
\item{theta}{variable to map angle to (\code{x} or \code{y})}

\item{start}{offset of starting point from 12 o'clock in radians}

\item{direction}{1, clockwise; -1, anticlockwise}

\item{bg}{raster for background image}
}
\description{
based on \code{\link[=coord_polar2]{coord_polar2()}}, but paths are connected linear. See the documentation
of \code{\link[=coord_polar2]{coord_polar2()}} for more Information and examples
}
\examples{
fn <- rOstluft.data::f("Zch_Stampfenbachstrasse_2010-2014.csv")
data <- rOstluft::read_airmo_csv(fn)
data <- rOstluft::rolf_to_openair(data)

funs <- list(
  "median",
  "mean",
  "q95" = ~ stats::quantile(., probs = 0.95)
)

data_summarized <- summary_wind(data, ws, wd, NOx, fun = funs,
  ws_cutfun = cut_number.fun(1)
)

# background map
bbox <- bbox_lv95(2683141, 1249040, 500)
bg <- get_stamen_map(bbox)

radar <- ggplot(data_summarized, aes(x = wd, y = NOx, color = stat, group = stat)) +
  geom_polygon(size = 1, fill = NA) +
  scale_color_viridis_d(end = 0.6) +
  scale_y_continuous(limits = c(0, NA), expand = c(0,0, 0, 0))

radar + coord_radar(start = - 22.5 / 180 * pi)

# add background map
radar + coord_radar(start = - 22.5 / 180 * pi, bg = bg)

# facets
radar + facet_wrap(vars(stat)) +
  coord_radar(start = - 22.5 / 180 * pi)
}
