% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutfuns.R
\name{cut_wd}
\alias{cut_wd}
\title{Cut wind direction into factor classes}
\usage{
cut_wd(wd, binwidth = 45, add_labels = TRUE, ...)
}
\arguments{
\item{wd}{numeric vector of wind directions in °}

\item{binwidth}{width for \code{\link[ggplot2:cut_width]{ggplot2::cut_width()}}}

\item{...}{passed to \code{\link[ggplot2:cut_width]{ggplot2::cut_width()}}}
}
\description{
Wraps \code{\link[ggplot2:cut_width]{ggplot2::cut_width()}} function with \code{width = binwidth. closed = "left", boundary = 0} as fixed
arguments
}
