% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutfuns.R
\name{cut_ws.fun}
\alias{cut_ws.fun}
\title{Partial function constructor to cut wind velocity (or others) into factor classes}
\usage{
cut_ws.fun(binwidth = 1, ws_max = NA, squish = TRUE, right = TRUE,
  reverse = FALSE)
}
\arguments{
\item{binwidth}{width of the bins}

\item{ws_max}{cut off wind speed at this maximum}

\item{squish}{If TRUE wind velocities greater than will be include as additional level ">ws_max"}

\item{right}{logical, indicating if the intervals should be closed on the right (and open on the left) or vice versa.}

\item{reverse}{reverse order of result. This is sometimes useful when plotting a factor.}
}
\value{
a partial \code{\link[=cut_ws]{cut_ws()}} function with ws as sole argument
}
\description{
Partial function constructor to cut wind velocity (or others) into factor classes
}
\seealso{
\code{\link[=cut_ws]{cut_ws()}}
}
