% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutfuns.R
\name{cut_ws.fun}
\alias{cut_ws.fun}
\title{wrapper to cut wind velocity (or others) into factor classes}
\usage{
cut_ws.fun(ws_binwidth = 1, ws_max = NA, ...)
}
\description{
wrapper to cut wind velocity (or others) into factor classes
}
