% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutfuns.R
\name{cut_ws}
\alias{cut_ws}
\title{Cut wind velocity (or others) into factor classes}
\usage{
cut_ws(ws, binwidth = 1, ws_max = NA, reverse = FALSE, ...)
}
\arguments{
\item{ws}{numeric vector of wind velocitiy}

\item{binwidth}{width of the bins}

\item{ws_max}{cut off wind speed at this maximum}

\item{reverse}{reverse order of result. This is sometimes useful when plotting a factor.}

\item{...}{passed onto \code{\link[base:cut]{base::cut()}}}
}
\description{
Wraps \code{\link[base:cut]{base::cut()}} with \code{breaks = seq(0, max(pmin(ws, ws_max, na.rm = TRUE), na.rm = TRUE), binwidth)} as fixed
argument
}
