% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{get_stamen_map}
\alias{get_stamen_map}
\title{Wrapper around ggmap::get_stamen_map}
\usage{
get_stamen_map(bbox, width = 640, zoom = "auto", color = "bw", ...)
}
\arguments{
\item{bbox}{list with items \code{left}, \code{bottom}, \code{right} and \code{top} as WSG84 coordinates. Additional \code{width}
in meters if \code{zoom = "auto"}. Use \code{\link[=bbox_lv95]{bbox_lv95()}} to create a compatible bounding box list}

\item{width}{of resulting map in pixels.}

\item{zoom}{zoomlevel from 0 to 18 or "auto"}

\item{color}{color or black-and-white. Changed default to "bw"}

\item{...}{forwarded to \code{\link[ggmap:get_googlemap]{ggmap::get_googlemap()}}}
}
\value{
a ggmap object (a classed raster object with a bounding box attribute)
}
\description{
The functions \code{\link[=get_googlemap]{get_googlemap()}} and \code{\link[=get_stamen_map]{get_stamen_map()}} tries to harmonize the calls to \code{\link[=ggmap]{ggmap()}}. The objective is
to get interchangeable functions with sensible defaults. For example automatic calculation of the zoom. The function
\code{\link[=bbox_lv95]{bbox_lv95()}} generates a bbox object compatible with both functions.
}
\section{Attribution}{

\itemize{
\item Toner and Terrain: Map tiles by Stamen Design, under CC BY 3.0. Data by OpenStreetMap, under ODbL.
\item Watercolor: Map tiles by Stamen Design, under CC BY 3.0. Data by OpenStreetMap, under CC BY SA.
}
}

\examples{
bb <- bbox_lv95(2683141, 1249040, 500) # site Zch_Stamfenbachstrasse
get_stamen_map(bb) \%>\% ggmap::ggmap()
}
