% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtraj.R
\name{ggtraj}
\alias{ggtraj}
\title{Plotting hysplit trajectory}
\usage{
ggtraj(data, mapping = aes(x = lon, y = lat, group = date, color =
  height), incr = -seq(24, 96, 24), lims = NULL,
  add_traj_labels = TRUE,
  color_scale = ggplot2::scale_color_gradient(name = "m agl."))
}
\arguments{
\item{data}{tibble containing hysplit trajectories, format preferably similar to that of the 'openair' package}

\item{mapping}{ggplot mapping, typically: aes(x = lon, y = lat, group = date, color = height)}

\item{incr}{sequence of hours to draw an marker on the trajetory. Default -seq(24,96,24); can be NULL - then, no increment markers are plotted}

\item{lims}{list with xlim and ylim items defining the map section. See \code{ggplot::coord_quickmap()}}

\item{add_traj_labels}{add text labels with date and time for every trajectory}

\item{color_scale}{ggplot2 color scale}
}
\value{
ggplot2 object
}
\description{
Plotting hysplit trajectory
}
\examples{
fn <- system.file("extdata", "2017_ZH-Kaserne-hysplit.rds", package = "rOstluft.data")
traj <- readRDS(fn)
traj <- dplyr::filter(traj, date < lubridate::ymd("2017-01-08"))
ggtraj(traj)

}
