% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggwindrose.R
\name{ggwindrose}
\alias{ggwindrose}
\title{ggplot wrapper to create a windrose (polar wind-bin frequency plot)}
\usage{
ggwindrose(data, mapping, ..., wd_binwidth = 45, ws_max = NA,
  fill_scale = scale_fill_viridis_d(), bg = NULL,
  wd_cutfun = cut_wd.fun(binwidth = 45),
  ws_cutfun = cut_ws.fun(ws_binwidth = 1, ws_max = NA))
}
\arguments{
\item{data}{tibble containing wind speed, wind direction and/or air pollutant concentration data}

\item{mapping}{ggplot2 mapping, e.g. aes(wd = wd, ws = ws); requires wd, ws}

\item{wd_binwidth}{numeric, binwidth for wind direction, typically %in% c(45, 22.5)}

\item{ws_max}{maximum wind speed cap; last wind speed bin contains all wind speeds > ws_max}

\item{fill_scale}{ggplot2 fill scale, e.g. scale_fill_gradientn(...)}

\item{bg}{raster map, e.g. ggmap object as plot background}

\item{wd_cutfun}{NULL or a function with which wind direction is cut into bins; per default (wd_cutfun == NULL): function(wd) wd_classes(wd, wd_binwidth = wd_binwidth)}

\item{ws_cutfun}{NULL or a function with which wind speed is cut into bins; per default (ws_cutfun == NULL): function(ws) ws_classes(ws, ws_max = ws_max)}
}
\value{
ggplot object
}
\description{
ggplot wrapper to create a windrose (polar wind-bin frequency plot)
}
\examples{
require(rOstluft)
require(rOstluft.data)
require(rOstluft.plot)
require(ggplot2)
require(dplyr)
require(openair)

df <-
  rOstluft::read_airmo_csv(system.file("extdata", "Zch_Stampfenbachstrasse_2010-2014.csv", package = "rOstluft.data", mustWork = TRUE)) \%>\%
  rOstluft::rolf_to_openair() \%>\%
  openair::cutData(date, type = "daylight")

ggwindrose(df, aes(ws = ws, wd = wd), cut_wd = cut_wd.fun(binwidth = 22.5), ws_cutfun = cut_ws.fun(ws_binwidth = 0.5, ws_max = 3))

ggwindrose(df, aes(ws = ws, wd = wd), wd_binwidth = 22.5, ws_cutfun = cut_ws.fun(ws_binwidth = 0.5, ws_max = 3)) +
  facet_wrap(daylight~.)

}
