% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggyearday.R
\name{ggyearday}
\alias{ggyearday}
\title{ggplot2 heatmap for diurnal-yearly time series}
\usage{
ggyearday(data, time, z, date_breaks = "1 month", date_labels = "\%b",
  ybreaks = seq(3, 21, 3), ylabels = format_sprintf("\%02d:00"),
  fill_scale = scale_fill_viridis_c(direction = -1, na.value = NA, option
  = "A"), ...)
}
\arguments{
\item{data}{a data.frame or tibble with input data (containing a POSIXct variable as time parameter).}

\item{time}{character string giving time variable name.}

\item{z}{character string giving z value variable name.}

\item{date_breaks}{character string as input for \code{\link[ggplot2:scale_x_date]{ggplot2::scale_x_date()}}, e.g. '1 month', defines date breaks on x-axis.}

\item{date_labels}{character string as input for \code{\link[ggplot2:scale_x_date]{ggplot2::scale_x_date()}}, formatter for date labels on x-axis.}

\item{ybreaks}{numeric vector, specifies y-axis breaks.}

\item{ylabels}{function, format function for y-axis labels.}

\item{fill_scale}{ggplot2 continuous fill scale, e.g. \code{\link[=scale_fill_gradient]{scale_fill_gradient()}}.}

\item{...}{other arguments passed on to \code{\link[ggplot2:geom_raster]{ggplot2::geom_raster()}}.}
}
\value{
ggplot
}
\description{
creates a heatmap with date on x-axis and time of day on y-axis; z values as colour scale.
}
\examples{
library(ggplot2)

fn <- rOstluft.data::f("Zch_Stampfenbachstrasse_2010-2014.csv")

df <-
  rOstluft::read_airmo_csv(fn) \%>\%
  rOstluft::rolf_to_openair() \%>\%
  dplyr::mutate(year = lubridate::year(date))

ggyearday(df, time = "date", z = "O3") +
  facet_wrap(year~., scales = "free_x", ncol = 1)

# data with outliers / extreme values => not very informative...
ggyearday(df, time = "date", z = "PM10") +
  facet_wrap(year~., scales = "free_x", ncol = 1)

# ...use a custom scale and squish the outliers / extreme values
fill_scale <- scale_fill_viridis_squished(breaks=c(0, 25, 50, 75), limits = c(0, 75),
                                          direction = -1, na.value = NA, option = "A")
ggyearday(df, time = "date", z = "PM10", fill_scale = fill_scale) +
  facet_wrap(year~., scales = "free_x", ncol = 1)


}
