% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_factor_funs.R
\name{midpoints}
\alias{midpoints}
\title{returns class midpoint-values for \verb{[x-y)}-style factor level}
\usage{
midpoints(x)
}
\arguments{
\item{x}{\verb{[x-y)}-style factor levels from cutting function}
}
\description{
returns class midpoint-values for \verb{[x-y)}-style factor level
}
