% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad.R
\name{pad_to_year}
\alias{pad_to_year}
\title{Pads data to complete year}
\usage{
pad_to_year(data, date_col, interval)
}
\arguments{
\item{data}{input data}

\item{date_col}{column containing date information}

\item{interval}{interval between two dates}
}
\value{
padded data
}
\description{
Generic padding function.
}
