% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcalendar.R
\name{plt_cal}
\alias{plt_cal}
\title{calendar plot}
\usage{
plt_cal(data, x = "starttime", z = "value",
  scale_fill = scale_fill_viridis_c(), size = 0.1, color = "white",
  ..., locale = Sys.getlocale("LC_TIME"))
}
\arguments{
\item{data}{input data}

\item{x}{date column as Date, POSIXct or Character}

\item{z}{value columns}

\item{scale_fill}{scale for tile fill. Default is for continous data. For discrete data one must be provided}

\item{size}{space between tiles}

\item{color}{color of space}

\item{...}{more options for \code{ggplot2::geom_tile()}}

\item{locale}{locale string for \code{lubridate::month()} and \code{lubridate::wday()}. see \code{Sys.getlocale()}}
}
\value{
ggplot2::ggplot object
}
\description{
calendar plot
}
