% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R, R/coord_polar2.R,
%   R/geom_polar_raster.R, R/rOstluft.plot-package.R, R/stat_filter.R
\docType{data}
\name{CalMonthBorder}
\alias{CalMonthBorder}
\alias{CoordPolar2}
\alias{GeomPolarRaster}
\alias{rOstluft.plot-ggproto}
\alias{StatFilter}
\title{All ggproto Objects}
\description{
All extended ggproto Objects
}
\seealso{
ggplot2::ggproto
}
\keyword{datasets}
\keyword{internal}
