% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rOstluft.plot-package.R
\docType{package}
\name{rOstluft.plot-package}
\alias{rOstluft.plot}
\alias{rOstluft.plot-package}
\title{rOstluft.plot: Ploting air quality data}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

This package provides function to quickly display data as diagrams.
}
\author{
\strong{Maintainer}: Thomas von Allmen \email{thomas.vonallmen@zuerich.ch}

Authors:
\itemize{
  \item Jörg Sintermann
  \item Noël Rederlechner
  \item Hanna Herich
}

Other contributors:
\itemize{
  \item Ostluft [copyright holder]
}

}
\keyword{internal}
