% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_summary_wind.R
\name{stat_summary_wind}
\alias{stat_summary_wind}
\title{Summarise z values over binned wind data.}
\usage{
stat_summary_wind(data = NULL, mapping = NULL, fun = "mean",
  fun.args = list(), nmin = 1, wd_cutfun = cut_wd.fun(binwidth = 45),
  wd_offset = 0, ws_cutfun = cut_ws.fun(binwidth = 1, ws_max = NA),
  groups = NULL, layer_args = list(geom = "polygon"),
  param_args = list())
}
\arguments{
\item{data}{The data to be displayed in this layer.
#' requires input data including at least three columns carrying information regarding:
\itemize{
\item wind direction (in °)
\item wind velocity
\item z-values (e.g. air pollutant concentration)
}}

\item{mapping}{ggplot2 mapping, e.g. aes(wd = wd, ws = ws, z = NOx); requires wd, ws, z}

\item{fun}{function or list of functions for summary.}

\item{fun.args}{a list of extra arguments to pass to fun.}

\item{nmin}{numeric, minimum number of values for fun, if n < nmin: NA is returned}

\item{wd_cutfun}{function, cut function for wind direction (to create bins)}

\item{wd_offset}{numeric, offset for wind_direction (in °) if groups == "wd"; bins are then calculated over (wd + wd_offset) %% 360}

\item{ws_cutfun}{function, cut function for wind speed}

\item{groups}{character string, can be NULL, "wd", ...}

\item{...}{other arguments passed on to layer(params = list(...)).}

\item{geom}{The geometric object to use display the data.}
}
\value{
ggplot2 layer
}
\description{
Binning is done by StatWind, so input data to stat_summary_wind() should be original unbinned data.
Depending on the groups argument, binning is either done 2-dimensional over cartesian u and v wind vectors
(calculated from input data; then, stat_summary_wind() yields results similar to openair::polarplot())
or,
1-dimensional over wind direction or wind velocity bins, respectively.
}
